Move getCurrencyIcon to CdnUris, rename to getCurrencyIconUris
