+++ src/components/scenes/Staking/StakeModifyScene.js
@@ -12,7 +12,7 @@ import { getDenominationFromCurrencyInfo, getDisplayDenomination } from '../../.
 import { useEffect, useState } from '../../../types/reactHooks.js'
 import { useSelector } from '../../../types/reactRedux'
 import type { NavigationProp, RouteProp } from '../../../types/routerTypes'
-import { getCurrencyIcon } from '../../../util/CurrencyInfoHelpers.js'
+import { getCurrencyIconUris } from '../../../util/CdnUris'
 import { getWalletFiat } from '../../../util/CurrencyWalletHelpers.js'
 import { getPolicyIconUris, getPolicyTitleName, getPositionAllocations, stakePlugin } from '../../../util/stakeUtils.js'
 import { zeroString } from '../../../util/utils.js'
@@ -246,7 +246,7 @@ export const StakeModifyScene = (props: Props) => {
   const renderChangeQuoteAmountTiles = modification => {
     return (
       <View style={styles.amountTilesContainer}>
-        <IconTile title={s.strings.wc_smartcontract_wallet} iconUri={getCurrencyIcon(currencyWallet.currencyInfo.pluginId).symbolImage}>
+        <IconTile title={s.strings.wc_smartcontract_wallet} iconUri={getCurrencyIconUris(currencyWallet.currencyInfo.pluginId).symbolImage}>
           <EdgeText>{currencyWallet.name}</EdgeText>
         </IconTile>
         {
